+++ src/Entity/MagicEntity.php
@@ -50,11 +50,15 @@ abstract class MagicEntity extends Entity implements JsonSerializable
     /**
      * PHP magic method.
      *
-     * @return array|null
+     * @return array
      */
-    public function __debugInfo(): ?array
+    public function __debugInfo(): array
     {
-        return $this->_hectorAttributes;
+        try {
+            return $this->_hectorAttributes + $this->getRelated()->__debugInfo();
+        } catch (OrmException) {
+            return $this->_hectorAttributes;
+        }
     }
 
     /////////////////////////